// src/app/communication-protocol/model/protocol-param.ts
export class ProtocolParam {
  commProtocolId: number;
  commProtocolParamGroupId: number;
  id: number;
  name: string;
  description: string;
  dataTypeId: number;
  engUnitId: number;
  isNull: boolean;
  defaultValue: string;
  uiComponentTypeId: number;
  displayOrder: string;
  status: string;
  refCommProtocolParamId: number;
  refCommProtocolParamValueId: number;
  dataTypeName:string;
  engUnitName: string;
  refProtocolParamValues: any;
}